+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Also pass the `model` instance when creating a new `DocumentList` instance
 * Fix `Document#remove` not finding the $model property
 * `DocumentList` instances can now be created without any arguments
+* Add `DocumentList#toSimpleArray()` which returns an array of simple objects with only the wanted fields
 
 ## 1.0.2 (2018-07-07)