+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `DocumentList` instances can now be created without any arguments
 * Add `DocumentList#toSimpleArray()` which returns an array of simple objects with only the wanted fields
 * Add `DocumentList.unDry` static method to the server-side class, because static methods don't get added by default yet
+* Fix `Document#conduit` causing an infinite loop
 
 ## 1.0.2 (2018-07-07)