+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `DocumentList#toSimpleArray()` which returns an array of simple objects with only the wanted fields
 * Add `DocumentList.unDry` static method to the server-side class, because static methods don't get added by default yet
 * Fix `Document#conduit` causing an infinite loop
+* Change `Model.cacheDuration` to `Model.cache_duration`
 
 ## 1.0.2 (2018-07-07)