+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Make `Model#save()` and `Model#findById()` also return Pledges
 * Make `Model#remove()`, `Document#remove()`, `Model#ensureIds()` return Pledges
 * Make MongoDataSource use `findOneAndDelete` to remove a record
+* Switched from `slug` dependency to `mollusc`, which is a fork with several fixes
 
 ## 1.0.2 (2018-07-07)