+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * When the `port` setting is `null`, a random port will be tried. All other falsy values will still not start the server.
 * `Alchemy#usePlugin(name, options)` now accepts a `path_to_plugin` option
 * `Alchemy#usePath()` will no longer load `test` folders
+* Fix paths equaling a prefix not working as you expect them to
 
 ## 1.0.2 (2018-07-07)