+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * `Alchemy#usePlugin(name, options)` now accepts a `path_to_plugin` option
 * `Alchemy#usePath()` will no longer load `test` folders
 * Fix paths equaling a prefix not working as you expect them to
+* Don't start socket.io when `settings.websockets` is false
 
 ## 1.0.2 (2018-07-07)