+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Fix paths equaling a prefix not working as you expect them to
 * Don't start socket.io when `settings.websockets` is false
 * Fix `Alchemy#broadcast(type, data)` not working
+* Add `Model#getBehaviour(name)`, which gets an existing instance on the current model
 
 ## 1.0.2 (2018-07-07)