+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.0.3 (WIP)
+## 1.0.3 (2018-07-12)
 
 * Update alchemy info page
 * `Schema#addField(name, type, options)` now allows you to directly pass a `Schema` instance as a type